build: Fix GBDK-2020 4.3 compatibility and produce ROM
- Fix platform flag: sm83:gbc -> sm83:gb (sm83:gbc not valid in GBDK 4.3)
- Add CGB-only flag via -Wl-yp0x143=0xC0
- Rename fill_rect -> map_fill in zone.c to avoid collision with GBDK macro
- Cast TILE_CREA_BASE + CREA_SPRITE_TILES to suppress overflow warnings
- Include built 32KB ROM: CreatureCollector.gbc

+++ game/src/battle.c
@@ -336,7 +336,7 @@ void battle_start(Creature *enemy, uint8_t boss_flag) {
     /* Generate and load procedural sprites */
     set_bkg_data(TILE_CREA_BASE, CREA_SPRITE_TILES,
                  sprite_gen_build(e_crea));
-    set_bkg_data(TILE_CREA_BASE + CREA_SPRITE_TILES, CREA_SPRITE_TILES,
+    set_bkg_data((uint8_t)(TILE_CREA_BASE + CREA_SPRITE_TILES), CREA_SPRITE_TILES,
                  sprite_gen_build(p_crea));
 
     /* Palette attributes for creature sprite areas */
@@ -560,7 +560,7 @@ uint8_t battle_update(void) {
                 p_last_stand = has_keystone(p_crea, NTYPE_LAST_STAND);
 
                 /* Reload player sprite */
-                set_bkg_data(TILE_CREA_BASE + CREA_SPRITE_TILES,
+                set_bkg_data((uint8_t)(TILE_CREA_BASE + CREA_SPRITE_TILES),
                              CREA_SPRITE_TILES,
                              sprite_gen_build(p_crea));
